+++ views.go
@@ -8,6 +8,10 @@ import (
 )
 
 func (s *Server) LoginViewHandler(w http.ResponseWriter, r *http.Request) {
-
+	// need to format with details.fqdn
 	fmt.Fprint(w, views.LoginView)
+}
+
+func (S *Server) CreateUserViewHandler(w http.ResponseWriter, r *http.Request) {
+	fmt.Fprint(w, views.AddUserView)
 }
