+++ views/addUser.go
@@ -1,4 +1,6 @@
-<!DOCTYPE html>
+package views
+
+var AddUserView string = `<!DOCTYPE html>
     <html lang="en"> 
         <head>
             <meta charset="UTF-8">
@@ -42,7 +44,7 @@
             <section class="section has-background-black-ter" style="height: 100vh;">
                 <div class="container">
                     <h1 class="title has-text-primary">Add User</h1>
-                    <form class="box has-background-black-ter" action="/adduser">
+                    <form class="box has-background-black-ter" id="addUserForm">
                         <div class="field">
                             <label class="label  has-text-primary" for="email">email:</label>
                             <div class="control">
@@ -66,14 +68,14 @@
             
                         <div class="field">
                             <div class="control">
-                                <button class="button is-primary" type="sumbit">Add User</button>
+                                <button class="button is-primary" type="sumbit">add</button>
                             </div>
                         </div>
                     </form>
                 </div>
             </section>
         </body>
-        <script type="module" src="./js/functions.js"></script>
+        <script type="module" src="./static/js/functions.js"></script>
     </html>
 
 
@@ -87,50 +89,4 @@ dropDownMenuLink.addEventListener('click',()=>{
 })
 
 </script>
-    <!-- <script>
-        document.getElementById('addUserForm').addEventListener('submit',function(event){
-            event.preventDefault();
-            const email = document.getElementById('email').value
-            const key = document.getElementById('key').value
-            const admin = document.getElementById('admin').value
-            const selectedInput = document.getElementById('selected').value
-            const selected = {};
-            selectedInput.split(',').array.forEach(pair => {
-                const [key,value] = pair.split(':');
-                if(key && value){
-                    selcted[key.trim()] = parseInt(value.trim(),10)
-                }
-            });
-            const userData = {
-                email,
-                key,
-                admin,
-                selected,
-            };
-            fetch('http://localhost:8080/adduser',{
-                method: 'POST',
-                headers:{
-                    'Content-Type': 'application/json',
-                    'Authorization': 'admin@aol.com:6953b699-9ee8-4cc2-bd2f-7259f49ff358'
-                },
-                body: JSON.stringify(userData),
-            })
-            .then(response => {
-                if (response.ok){
-                    return response.json();
-                }
-                throw new Error('Network response was not ok.');
-            })
-            .then(data => {
-                console.log('User added: ',data);
-                alert('User added successfully!')
-            })    
-            .catch(error=> {
-                console.error('There was a problem with the fetch operation',error);
-            });
-
-        });
-
-
-
-    </script> -->
+`
